:art: Explanation page apis (#80)
* ICE, Fairml, Skater example usage

* SHAP library tested

* try rulefitr

* Skater decision tree different visualiztion

* SHAP library api is ready

* SHAP, Skater, ICE and decode api is ready

* Clean project and add new requirements to requirements.txt

* SHAP API bug fix for regression

* requirements.txt bug fix

* requirements.txt bug fix

* requirements.txt bug fix

* requirements.txt bug fix

* requirements.txt bug fix

* requirements.txt bug fix

* requirements.txt bug fix

* test fix

* requirements.txt version changes

* requirements.txt version changes

* requirements.txt version changes

* test_refactoring bug fix

* get train and test log from split APIs ready

* Explanation page libraries done

* Explanation page libraries done

* Get train and test logs based on split API is ready

* Failed test fix

* Failed test fix

* Failed test fix

* Fix tests

* Fix tests

* Fixing tests in test_refactoring.py

* LIME api fix for regression model

* Tests for get train and test log files

* Code cleaning

* Update test_views.py

* Code cleaning

* Update lime_wrapper.py

* Update requirements.txt

* CfFeedback and retrain api

* Add prettyjson to requirements.txt

* Small fix

* Small fix

* Small fix

* Small fix

* Small fix

* Small fix

+++ src/explanation/shap_wrapper.py
@@ -36,5 +36,7 @@ def explain(shap_exp: Explanation, training_df, test_df, explanation_target):
     f = open(name, "r")
     response = f.read()
     os.remove(name)
-    os.remove(name.split('.svg')[0])
+    if os.path.isfile(name.split('.svg')[0]):
+        os.remove(name.split('.svg')[0])
+
     return response
